+++ Invoke-Environment.ps1
@@ -5,14 +5,18 @@
 	Author: Roman Kuzmin (inspired by Lee Holmes's Invoke-CmdScript.ps1)
 
 .Description
-	It invokes the specified command (normally a configuration batch file) and
-	imports its environment variables to the current PowerShell environment.
+	It invokes the specified CMD command and imports its environment variables
+	to the current PowerShell session. The command is normally a configuration
+	batch file without arguments or very simple arguments, e.g. arguments with
+	spaces are troublesome.
+
+	Command output is discarded. Use the switch Output in order to receive it.
+	$LASTEXITCODE can be examined unless the switch Force is specified.
 
 .Parameter Command
 		Specifies a command being invoked, e.g. a configuration batch file.
 		This string is the whole command text passed in cmd /c. Do not use
-		redirection operators in it, they are used by the script. Output is
-		discarded by default. Use the switch Output in order to receive it.
+		redirection operators, they are used by the script.
 .Parameter Output
 		Tells to collect and return the command output.
 .Parameter Force
@@ -45,7 +49,7 @@ param
 $stream = if ($Output) { ($temp = [IO.Path]::GetTempFileName()) } else { 'nul' }
 $operator = if ($Force) {'&'} else {'&&'}
 
-foreach($_ in cmd /c "$Command > `"$stream`" 2>&1 $operator set") {
+foreach($_ in cmd /c " $Command > `"$stream`" 2>&1 $operator SET") {
 	if ($_ -match '^([^=]+)=(.*)') {
 		[System.Environment]::SetEnvironmentVariable($matches[1], $matches[2])
 	}